Adds expect=pass|fail and print=true|false attributes.
The expect attribute can be applied to Test or Defaults and says whether or not the test is expected to pass.
The default value for expect is "pass", which means tests are expected to pass unless explicitly stated otherwise.

The print attribute is a boolean that controls whether or not fileTester prints anything to stdout. If it is set to
true, then fileTester will print messages of the form: testname: [result] to stdout for each test, as well as a final
line for each test file saying that it finished running the tests in that file. False will prevent messages from
begin printed to stdout, but the information will still be written to the log file.