This commit adds server-side encryption using keys derived from a keyczar HMAC Signer; key generation is determisitic with respect to the HMAC and its key, so the server always rederives the same keys, hence can read its files.
On encryption, the server writes a metadata file containing the IV, the hmac,
the name, and the size, and it HMACs this data for decryption time.

The current implementation here uses AES-CBC and HMAC-SHA1 and matches the
original CloudProxy in doing streaming encryption instead of breaking up
encrypted files into blocks.