A rewritten version of the kvm/tcservice path.
This version now gets a guest virtual address from the hypercall, then uses this
to copy data from the guest pages to the tcService queues in the host kernel.
This commit also removes support for multiple tcioDD devices, since there will
only ever be one in the host kernel.

We use the vcpu->vcpu_id for the process id, but this should probably change.

The script copy_to_kernel.sh <kernel dir> takes in a host kernel directory (default
/usr/src/linux-3.10.1 and copies the files into this directory. It doesn't yet
handle patching the files that need to be patched, so be warned that this needs
to be done as well.

Also, this patch removes all dependencies on the fileProxy code and removes the
cflags annotation from the tcioDD makefile.