Rewrite of CloudClient and CloudServer to use SSL objects instead of BIO objects. This also fixes a double-free bug that was in CloudClient: it was freeing an X.509 self certificate that was owned by the SSL_CTX and was also freed in the ScopedSSLCtx destructor.
This commit generalizes the OpenTCPSocket method to accept a host and port
combination and updates the applications to use this feature. It also adds
SO_REUSEADDR to the socket in OpenTCPSocket so that it can take over abandoned
sockets that are in state TIME_WAIT (this allows the tests to run in quick
succession on the same port).