TPM and fileproxy fixes, DebugPrint functions
TCCA now properly strips off the tpm quote when condensing an
attestation chain, and LinuxTao only checks to be sure the same key
was used for both the original chain and condensed attestation.

DebugString() for pretty-printing an Attestation or Statement.

Minor fixes to make fileproxy tests work.

Typos and error messages.