Support for shared secrets.
Hosted program can call Tao::GetSharedSecret() to obtain a random
secret from the host Tao. A policy controls which clients can access
the same shared secret, e.g. just the calling child, all children,
etc. The secret is a deterministic function of the size and policy
specified. This can be used to eliminate the Seal and Unseal
interfaces. (The reverse is not true, though if the encryption
operation for seal was deterministic, we could probably use it to
implement GetSharedSecret.)