go auth and attestation code (partial)
Package auth contains code for parsing and printing auth-related
objects: Prin (for principals), Pred (for predicates), etc. There is
code for converting string to Prin and back, and the conversions play
nicely with Printf and Scanf.

attestation.go contains the beginnings of code for verifying
attestations.

All unit tests pass (c++ and go), and the go demo works on top of the
c++ linux host in TCP and TLS modes (Tao auth mode is unfinished).