Changes to take into account review and lint.
This commit makes many changes to account for the comments in host.diff as well
as comments returned by golint and go vet. All the change are fairly minor.
However, we also move tao_rpc.pb.go and replace the link we used to have for it.
This is because we're starting to diverge from the C++ implementation; this
commit removes the error field, since go rpc handles errors directly.