Use go-friendly rpc encoding in C++ and go
Go's net/rpc API is more friendly if the codec uses two messages for
each request and for each response: one header message carrying a
sequence number, operation (as a string), and optional error message;
and a separate message for the body.

This patch uses the two-part style in both C++ and go. The go code is
much simpler. The C++ code, which is in the process of being
abandoned, is barely more complicated.

Note that this still uses the protobuf codec for all messages, both
headers and bodies. This is sensible because protobuf supports
multiple languages easily and it has a compact, simple, and
well-documented encoding.