Changed unpack/pack structure to use reflection.
This is along the lines of suggestions made by Kevin in the tpm_seal.diff code
review.

This version of the code more accurately handles the TPM data structures. In
particular, it seals and unseals using the TPM-defined tpmStoredData structure,
though it returns this from Seal an opaque blob rather than the internal data
structure. It also handles the length of a slice directly in the encoding and
decoding; it does this by doing a recursive encoding that reflects on the type
of elements being encoded and explicitly reading and writing the size of slices.
This means that encoded data structures can now explicitly use slices.