Commit by Kevin Walsh: Add quantification for auth.
This commit adds support for "forall" and "exists" statements in the
authorization language and AST. This is needed to set up clean ways of
describing the authorization of TPMs and OSes and programs in the standard
CloudProxy setup in datalog.

This commit changes some ACL and TPMTao code to port it to the new auth setup,
especially with respect to auth.Bytes and changes to some elements that used to
be exclusively auth.Prin but have been generalized to auth.Term.