Add hybrid sealed data and fix the install script to support TPM.
This commit adds support for TPM-based tests using install.sh -tpm.
It adds a new hybrid sealing scheme in TPMTao to overcome a limit in the number
of bytes that can be sealed in the hardware TPM; to do this, it creates a new
key (using Crypter in keys.go), encrypts the data using this key, and seals the
serialized key structure as the sealed data. Unseal reverses the operation.

This version also changes the host to not try to extend the name of the Tao in
the Stacked LinuxHost when the Tao is TPMTao. This is OK because TPMTao already
extends its name with the PCR values, which characterize the Linux OS in
question (or at least, they do if we boot from initrd).