Fix basic datalog translation in DatalogGuard.
This commit fixes a bug in ReloadIfModified: the datalog engine was getting
cleared, but the assert method was skipping rules that already existed in the
database. So, after a ReloadIfModified, the rule set was right, but the datalog
engine was out of sync. This also adds a regression test for this case.