Get datalog to work over the TPM.
This commit adds support for datalog authorization policies over the TPM and
gets the basic policy in install.sh to work. The problem with the version of the
subprin/3 principal before this commit was that it wasn't handling the quoted
strings correctly. This version outputs quoted strings as well as parsing
quoted-string principals in subprin/3.