Add support for labeled Tao channel types.
This commit adds support for a new shell variable: GOOGLE_HOST_TAO_TYPE. It also
adds support for two values for this variable "tpm" and "pipe". These correspond
to the two currently support ways for a program to talk to a parent Tao: either
it creates a TPMTao and uses, e.g., /dev/tpm0, or it is given a specification of
a pair of pipes that should have been added to its file descriptors at fork
time, and it uses those pipes for communication.

The next option to be supported will be "unix" for communication over a unix
socket, since this will be the easiest way to support Tao channels into docker.