Add interfaces for hosted programs.
This commit adds an interface for HostedProgram: this generalizes the
os/exec.Cmd and the github.com/docker/docker/daemon.Container interfaces to
support starting and stopping processes and containers. This also adds an
interface for hosted-program factories. This generalizes the LinuxProcessFactory
type and will soon be generalized to add support for docker containers.