Add support for Docker containers.
This commit adds a new factory: the LinuxDockerContainerFactory, and adds
support in install.sh for starting hosted programs under Docker containers. The
container is based on the empty Docker "scratch" container, so the hash of the
tarball that defines the image uniquely defines its contents under Docker. So,
the subprincipal for a Docker container is Container([bytes]), where [bytes] is
the SHA-256 hash of the tarball.

In install.sh new auth rules for the datalog guard state that if a program hash
is trusted, a container hash is trusted, and the container contains the program,
then this container can act as a MemberProgram. This hooks containers back into
the original rules about member programs and authorization to execute.

Containers get Unix domain sockets hooked in from the host file system and use
them for Tao communication with the parent Tao.