Add initial LinuxHost/CoreOS support.
This commit adds support for running LinuxHost as a privileged Docker container
on a CoreOS virtual machine under QEMU/KVM using the KVMCoreOSFactory. It also
adds support for channels that act exactly like regular files in the filesystem.
This is the kind of channel that is created on the guest in /dev/virtio-ports
using virtio-serial under QEMU/KVM.