Add a code review for changes to fileproxy.
This commit adds a code review for fixes that update fileproxy to match the
style of the rest of the Go cloudproxy implementation. These changes also
refactor much of the code to make better use of the MessageStream implementation
and to reuse as much of the keys.go implementation as possible.