Add a script to run KVM/CoreOS Tao with apps.
This commit adds a script that runs a hypervisor Tao, then starts a virtual
machine with KVM/CoreOS, sets up the guest-layer linux_host, then runs the demo
server and client on it, collects their logs, and cleans up. The KVM/CoreOS
factory now takes a path to use for the linux_host files and a port for SSH.
This is so that these can be specified by tao_launch, hence are known to the
caller.