Fix references to proto and fix run_kvm.sh
This commit replaces references to code.google.com/p/goprotobuf/proto with
github.com/golang/protobuf/proto. It also changes the KVM/CoreOS
HostedProgramFactory to create a special RSA key to use for setup. It adds this
key to the authorized hosts along with the keys passed at the command line.