Update tao_admin to set up domains better.
This commit changes the tao_admin program to independently set up domains in
ways that can easily be scripted. It also adds some sample domain templates and
changes code that was relying on the old gcfg format. It also adds a simple
script that sets up a new datalog domain for testing with a root domain based
on a key.