Add support for running process and docker tests.
This commit changes the way the process and docker scripts work and generalizes
the set_up_domain.sh script to handle predicates for hosts, programs, and
containers in a configurable way. The run_process.sh and run_docker.sh programs
work. This also fixes build_standalone.sh, which was broken by a change in go
1.4 that builds static binaries in a new way.