Fix the new TPM support to allow Key principals.
This commit changes the new TPM scripts to work if a TPM isn't present.
In particular, it adds guards to only add a TPM principal if the
appropriate files are available. And it fixes a bug in the TPM support
in tao_admin.go to return the errors in the case of errors in finding
the TPM files.