Add Directive protocol buffer for the mixnet.
Router.HandleProxy() reads a cell from a proxy. The first byte of the
cell determines whether the proxy is sending a message or a directive.
If it is a message, the next 8 bytes encode the length of the whole
message and hence the number of cells it expects from the proxy. If
it is a directive, the cell is unpaded and unmarshaled. The length of
a marshaled Directive must be less than that of a cell.