Refactor HostedProgramFactory and its params.
This is a complete refactoring of hosted program factory and the
calls for starting hosted programs. All parameters for controlling the
creation of a hosted process are now in a single HostedProgramSpec
struct. This mirrors the style of os/exec.Cmd, somewhat.

The factory now has a single NewHostedProgram method, taking spec and
returning a HostedProgram. There are factory-specific structs that
satisfy the HostedProgram interface:
 HostedProcess for LinuxProcessFactory
 DockerContainer for LinuxDockerContainerFactory
 KvmCoreOSContainer for LinuxKVMCoreOSFactory

On a HostedProgram, the host can call Subprin() to get the subprin,
then Start(), Kill(), etc. During creation of each struct, all
necessary toc-tou copying and hashing is done. Some of the final setup
is done in Start(). There is also a Cleanup() which attempts to be a
little better about removing temp files.