Change default args behavior of linux_host
Change a few of the arguments for linux_host. If given -config_path,
it will expect that to exist and will use it as the config. If given a
new -temp_trivial_domain option, it will create a temporary trivial
domain configuration (this option is used by fileproxy/run.sh). If
given neither, it tries env var $TAO_DOMAIN_CONFIG.

Fix fileproxy/run.sh to match this usage. Also in that script, use
gowhich so GOPATH can have multiple paths.

fileproxy/run.sh is still broken, but not clear how long it has been
broken.