tao_launch proxies stdio and signals
The tao_launch front-end app now proxies stdin/stdout/stderr back and
forth to the hosted program that was launched, and it waits until that
hosted program exits before exiting itself. There is now a -disown
option to do simple redirection and quit immediately instead, and a
-daemon option to pipe all stdio to /dev/null and quit immediately.

Waiting for exit requires support by the tao host admin rpc interface,
because it is tricky to wait for a non-child process.

Regenerated the protobuf files, which seems to have added an init
block..