Proxy client/server protocol over mixnet.
The mixnet will proxy traffic following a specific protocol: the client
dials the server then sends a message and waits for a response any
number of times. This is accomplished with an additional Queue structure
in RouterContext called prevQueue. The router queues messages as well as
channels for replies. When the router reads a response on a channel, it
divides it into cells and enquuees prevQueue for sending the respsonse
back to the client.