Implement SOCKS proxy and end-to-end testing.
ProxyContext now binds a socket to a port that exposes the SOCKS5
protocol to clients. This is meant to allow routing arbitrary
protocols over mixnet. This pull request improves the coverage of
testing (including timeouts) and an end-to-end test which runs a single
router and destination with many client/proxies. test_mixnet.sh has has
been modified to test with netcat.