Add support for registering arbitrary Tao channels.
This commit adds a Register call that takes a name and a function and
records them for later use a Tao channels. If the name is given in the
Tao config and there is a registered function with the same name, then
this function is used to generate the Tao host channel.