principals use a hash of key rather than full key
Previously, principal names (auth.Prin) encoded a serialized public
key. For example:
  key([..encoded key bytes...] ]).TrivialGuard("Liberal")
This served two purposes: it uniquely identifies the principal;
and it allows anyone holding a name to verify signatures produced by
that principal. For example, a tao.Attestation structure includes a
statement, the signature bytes, and the name of the principal that
signed the statement. To verify the attestation, one would extract the
public key from the principal name, then use that key to verify the
signature bytes against the statement. However, principal names are
somewhat large in this scheme.

This commit shortens principal names by encoding only the hash of the
serialized public key, so principals now appear as:
  key([...sha256 hash...]).TrivialGuard("Liberal")
This still uniquely identifies the principal. To verify a signature,
however, will require obtaining the actual un-hashed public key
through by some other means. In some cases, keys may already known or
are shared through other channels. In other cases, keys will need to
be distributed along with principal names.

For simplicity, this commit changes tao.Attestation so that it
includes the signer's full public key instead of the signer's
principal name. The principal name, encoding only a hash, is not
useful for verification of the signature bytes. The public key can be
used to verify the signature, then it can be hashed to derive the
signer's principal name. This change to tao.Attestation makes the
shortening of principal names mostly transparent to existing code.

All tests in tao/... compile and pass.

Demo using run/scripts/set_up_domains.sh and
run/scripts/run_processes.sh works.

Code in the following directories is now broken and should be
rewritten:
  go/apps/simpleexample/SimpleDomainService
  go/apps/fileproxy/keynegoserver