Add support for TPM2
This commit threads the TPM 2.0 code into the existing host/domain
infrastructure. It passes the configuration through the code down to
TPM2Tao, and it can be used to init a domain, then init a host.

However, this TPM 2.0 generates a random key each time; it needs to be
completed by setting up keys in files and having the code read it when
the TPM2Tao is instantiated.